+++ libraries/waspy/wasdi/CHANGELOG.md
@@ -3,6 +3,12 @@
 WASDI python programming library CHANGELOG.md, according to the 
 [standard proposed by keepchangelog](https://keepachangelog.com/)
 
+## [0.8.6.0] - 2024-02-22
+### Added
+- sleep parameter to control the waiting time while polling api
+### Fixed
+- moved import methods in the POST version
+
 ## [0.8.5.7] - 2023-12-20
 ### Added
 - enable auto donwnload from S3 Volumes